requirements.txt
absl-py>=2.0.0
Flask>=2.0.0
gunicorn>=20.0.0
h5py>=3.0.0
Jinja2>=3.0.0
keras>=2.9.0
numpy>=1.20.0
tensorflow-cpu>=2.9.0
Werkzeug>=2.0.0
